% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunction.R
\name{AnalysisFunction}
\alias{AnalysisFunction}
\title{Analyze a vector (by adding white noise!)}
\usage{
AnalysisFunction(x, training = T)
}
\value{
A numeric vector that has been analyzed.
}
\description{
This function analyzes a numeric vector. If \code{training = TRUE} (the default), the function adds random noise Z, with
\eqn{Z \sim Uniform(-1, 1)}.
}
\examples{
x <- 1:10

# Analyze a vector without added noise (boring)
y1 <- AnalysisFunction(x, training = F)
plot(x, y1) # no noise

# Analyze a vector with added noise (fun)
y2 <- AnalysisFunction(x, training = T)
plot(x, y2) # added noise

}
