% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataAsset.R
\name{GetDataAsset}
\alias{GetDataAsset}
\title{Download Lalonde's data from Dropbox}
\source{
LaLonde, R. J. (1986). Evaluating the econometric evaluations of training programs with experimental data. The American Economic Review, 604-620.
}
\usage{
GetDataAsset()
}
\value{
A data frame with Lalonde's NSW data (614 rows and 10 columns):
\describe{
\item{treat}{Treatment assignment (dummy)}
\item{age}{Participant's age (years)}
\item{educ}{Participant's years of schooling (years)}
\item{black}{Whether the participant was black (dummy)}
\item{hispan}{Whether the participant was hispanic (dummy)}
\item{married}{Whether the participant was married (dummy)}
\item{nodegree}{Whether the participant has no high school degree (dummy)}
\item{re74, re75, re78}{Participant's income in 1974, 1975, and 1978 (dollars)}
}
}
\description{
This function downloads Lalonde's NSW data from Dropbox, outputting it as \code{data.frame} object.
}
\examples{

# Download the data and save it as a data frame:
PackageDataAsset <- GetDataAsset()

}
